==> Script/regional_model/05_Extrapolation_forSpinUP_regional_NEW.Rmd <==
---
title: "Reconstruct effort for spin-up (1850-1950)"
author: "Camilla Novaglio (revised by Gage Clawson)"
date: "November 2024"
output: 
  html_document:
    toc: yes
    toc_float: yes
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE, message=FALSE)
```

## check global file for notes. 
## Set environment 

```{r environment, echo = FALSE}

rm(list=ls())

library(tidyverse)
library(dtplyr)
library(tictoc)
library(data.table)
library(mgcv)
library(gridExtra)

select<-dplyr::select

gem_dir <- file.path("/home/ubuntu/gem") # change to /rd/gem if that is the filepath for you 

```

## Effort data 

```{r effort, echo = FALSE}

effort <- fread(file.path(gem_dir, "private/users/yannickr/effort_histsoc_1950_2017_regional_model.csv"))

# Remove Unknown_EEZ before calcualting spin-up. See global file for note 
effort <- effort %>% 
  filter(FGroup != "Unknown_EEZ", !is.na(region))

```

## Catch data 

```{r catch, echo = FALSE}

catch <- fread(file.path(gem_dir, "private/users/yannickr/catch_histsoc_1869_2017_regional_models.csv"))
catch <- catch %>% 
  filter(!is.na(region))

# calculate catch as sum of reported and IUU, discards are generally not considered
catch$catch_tot <- catch$Reported + catch$IUU

```

## Define plot theme 

```{r theme, echo = FALSE}

my_theme <- theme_bw()+
  theme(text = element_text(size = 10), # this should be overwritten by the below
        plot.title = element_text(size = 10),
        axis.title = element_text(size = 9),
        axis.text = element_text(size = 8),
        legend.title=element_text(size = 9), 
        legend.text=element_text(size = 8),
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(), 
        legend.key.size = unit(0.1, "cm")) 
  
```

## Extrapolate total effort by LME function

```{r extrapolate, echo = FALSE}

toKeep <- sort(unique(effort$region))

effort_extrapolate <- function(toKeep, kappa, war_param, war_decision, reference){
  
  # # # trial
  # toKeep="Baltic Sea EwE"
  # kappa=4
  # war_param = "include"
  # war_decision = "no-add"
  # reference = "1945-1955"

  effort_extr <- effort %>% 
    filter(region %in% toKeep) %>% 
    group_by(Year) %>% 
    summarise(NomActive = sum(NomActive, na.rm = TRUE)) %>% 
    ungroup()

  maxYearEffort <- max(effort_extr$Year) 
  
  catch_all <- catch %>% 
    filter(region %in% toKeep) %>% 
    group_by(Year) %>% 
    summarise(catch_tot = sum(catch_tot, na.rm = TRUE))
  
  ## IF ZERO ----
  # if(dim(effort_extr)[1] !=0){ 
  
  ## IF HISTORICAL catch exist ----
  if(dim(filter(catch_all, Year < 1950))[1] >= 10) {  
    df = catch_all
    var = "catch_tot" 
    chechDataUsed = "catch"
  }else{
      df = effort_extr
      var = "NomActive"
      chechDataUsed = "effort"
      }
  
  # find the pick in data and filter out everything after the pick  
  maxVar <- max(df[[var]])  
  maxVar <- filter(df, eval(as.name(var)) == maxVar)
  toAddBack <- filter(df, Year > maxVar$Year)
  df <- filter(df, Year <= maxVar$Year)
    
  # filter out war years: 1914-18 and 39-45
  year <- sort(unique(df$Year))
  if (year[[1]]<=1918){
    war <- df %>% 
      filter(Year %in% c(1914:1918, 1939:1945)) 
    warToRemove <- war$Year
    }else if (year[[1]]<=1945){
      war <- df %>% 
        filter(Year %in% c(1939:1945)) 
      warToRemove <- war$Year
    }

  if(year[[1]] <= 1918 | year[[1]] <= 1945 & war_param == "exclude"){
    df <- filter(df, !Year %in% warToRemove)
    }
    
  # fit the gam 
  gam_model <- gam(eval(as.name(var)) ~ s(Year, k=kappa), data=df, family=gaussian(link="log"))
  df$gam <- round(predict(gam_model,newdata = df,type='response'))

  # # check the gam
  # ylim = c(min(df$gam), max(df$gam))
  # plot(df$Year,df[[var]],  ylim=ylim)
  # par(new=T)
  # plot(df$Year,df$gam,col="red",type='l', ylim=ylim)

  # predict backwards
  newd <- data.frame(Year = seq(1861, maxVar$Year))
  newd$pred <- predict.gam(gam_model,newd,type='response') 
  newd$pred <- ifelse(newd$pred<=0, 0, newd$pred)
  
  # all data
  df <- df %>%
    full_join(newd, by = "Year") %>% 
    dplyr::select(-gam)
    
  if(var == "catch_tot"){
    
    # 1. calculate rate of change
    
    if(reference == "1950"){
      ref <- df %>% 
        filter(Year == 1950)
      ref <- as.numeric(ref$pred)
    }else{
      # 1B. calculate rate of change based on mean 1945-1955 years 
      ref <- df %>% 
        filter(Year >= 1945, Year <= 1955)
      ref <- as.numeric(mean(ref$pred, na.rm = TRUE))
    }
    
    # re-add war years for calculation of effort (if these were available). Add them to pred, because pred <1950 is what you are basing your calculation of effort on.
    if(war_param == "exclude" & war_decision == "re-add"){
      
    if (year[[1]]<=1914 | year[[1]]<=1939){
      df<-df %>% 
        full_join(war %>% rename(var_war = as.name(var)), by = "Year") %>% 
        mutate(pred = case_when(!is.na(var_war) ~ var_war,
                                is.na(var_war) ~ pred)) %>% 
        select(-var_war)
    }
    }
      
    rate_of_change <- df %>% 
      arrange(Year) %>% 
      mutate(pred = as.numeric(pred)) %>% 
      mutate(rate = 100 * (pred - ref)/ref) %>% 
      mutate(Trial_back = ref + (rate/100)*ref) # only to check  
  
    # check 
    # filter(rate_of_change, Year >= 1950, Year <= 1960)

    # 2. extend effort back using rate of change in catch 
    
    if(reference == "1950"){
      ref<-effort_extr %>% 
        filter(Year == 1950)
      ref<-as.numeric(ref$NomActive)
    }else{
      # 2B. extend effort back using rate of change in catch using decadal mean 
      ref <- effort_extr %>% 
        filter(Year >= 1945, Year <=1955)
      ref<-as.numeric(mean(ref$NomActive, na.rm = TRUE))
    }
  
    effort_reconstruct <- effort_extr %>% 
      full_join(rate_of_change, by = "Year") %>% 
      mutate(NomActive_catch_based = case_when(Year >= 1950 ~ NomActive,
                                               Year < 1950 ~ ref + (rate/100)*ref))
    
    # # check
    # ggplot(data = effort_reconstruct, aes(x = Year, y = NomActive_catch_based)) +
    #   geom_point()
    # ggplot(data = effort_reconstruct, aes(x = Year, y = pred)) +
    #   geom_point()

    # build effort_all
    effort_all <- effort_reconstruct %>% 
      dplyr::select(Year, NomActive_catch_based) %>% 
      rename(NomActive = NomActive_catch_based) 
    
    # consider only recent years that are in effort and not in catch 
    effort_all <- effort_all %>% 
      filter(Year <= maxYearEffort)
    
    }else{ # END var = cacth_tot; beginning of var = NomActive 
      
      # build effort based on gam on effort. 
      df <- df %>% 
        mutate(NomActive = ifelse(Year < 1950, pred, NomActive)) 
      
      } # end difference between var 
        
    # FINAL STEPS ----
    
    # 1. add missing data 
    
    # add back war time in raw catches - only if you had excluded them above
    # WARNING - CHECK THIS 
    if(war_param == "exclude"){
  
    if (year[[1]]<=1918 | year[[1]]<=1945){
      df <- df %>% 
      full_join(war %>% rename(var_war = as.name(var)), by = "Year")
    
      if(var == "catch_tot"){
        df <- df %>% 
          mutate(catch_tot = case_when(!is.na(var_war) ~ var_war,
                                       is.na(var_war) ~ eval(as.name(var)))) %>% 
          select(-var_war)
      }
      # }else{ 
      #   # is this needed? was years are only excluded on catches, not effort. 
      #   df<-df %>% 
      #     mutate(NomActive = case_when(!is.na(var_war) ~ var_war,
      #                                  is.na(var_war) ~ eval(as.name(var)))) %>% 
      #     select(-var_war)
      #   
      # }
    }
    }
    
    # add back data after pick in both catches and effort
    if (dim(toAddBack)[1]!=0){
      df <- df %>% 
        full_join(toAddBack) 
      }
  
    # define df as either catch_all or effort_all
    if(var == "catch_tot"){
      catch_all2 <- df # version 2 otherwise cannot # highlight data that was not used in GAM # as catch_all is overwritten
    }else{
      effort_all <- df
      catch_all2 <- catch_all 
    }
    
    # 2. merge and plot 
    final_df <- effort_all %>% 
      full_join(catch_all2, by = "Year")
    
    # add 1841-1860 at 1861 fixed values
    trial <- final_df %>% 
      filter(Year == 1861) 
  
    add_df <- data.frame(Year = seq(1841, 1860), NomActive = rep(trial$NomActive, 20))
  
    final_df <- final_df %>% 
      full_join(add_df)

    # WARNING - check Warning with gather, use pivot_longer instead
    final_df <- final_df %>% 
      pivot_longer(cols = -Year, names_to = "Type", values_to = "Value") 
 
    # final_df<-final_df %>% 
    #   gather("Type", "Value", -Year) 
  
    # grouping for plot below - these are the titles of the two stacked plots 
    if(var == "catch_tot"){
      final_df <- final_df %>% 
        mutate(group = ifelse(Type == "NomActive", "Effort (nominal, DkW)", "Catch (tonnes)") )
      
    }else(
      final_df <- final_df %>% 
        mutate(group = ifelse(Type == "catch_tot", "Catch (tonnes)", "Effort (nominal, DkW)") )
    )
    
    # highlight data that was not used in GAM 
    if(dim(filter(catch_all, Year <1950))[1] >=10 & war_param == "exclude" & war_decision != "re-add") { 
      final_df <- final_df %>% 
        mutate(Considered = ifelse(Year %in% c(1914:1918, 1939:1945), "No", "Yes"))
    }else{ 
      final_df <-final_df %>% 
        mutate(Considered = "Yes")
    }
    
    if(dim(toAddBack)[1]!=0){
      final_df <- final_df %>% 
        mutate(Considered = ifelse(Year > maxVar$Year, "No", Considered))
    }
    
    final_df_plot <- final_df %>% 
      rename(Data = Type) %>% 
      mutate(Data = ifelse(Data == "pred", "GAM predictions", "Reconstructed values"),
             Considered = as.factor(Considered))
    
    # make "yes" always a dot and never a square 
    scale_fill_chris <- function(...){
    ggplot2:::manual_scale(
        'shape', 
        values = setNames(c(16,15), c("Yes", "No")), 
        ...
    )
    }
    
    # create data for annotation 
    yMax = max(filter(final_df, Type == "NomActive")$Value, na.rm = TRUE)
    # ann_text <- data.frame(Year = c(1841+10,1861+15,1961+15), Value = rep(yMax,3), lab = c("Spin-up","Transition","Experiment"), group = rep("Effort (nominal, DkW)", 3), Data = rep("Reconstructed values", 3), Considered = rep("Yes", 3))
    
    ann_text <- data.frame(Year = c(1841+10,1961+15), Value = rep(yMax,2), lab = c("Transition","Experiment"), group = rep("Effort (nominal, DkW)", 2), Data = rep("Reconstructed values", 2), Considered = rep("Yes", 2))
    
    plot <- ggplot(data = final_df_plot, aes(x = Year, y = Value, group = Data, color = Data, shape=Considered)) +
      ggtitle(paste(toKeep, sep = " "))+
      facet_wrap(~group, scales = "free", nrow = 2)+
      annotate("rect",xmin=1841, xmax=1960, ymin=0, ymax=Inf, fill = "#b2e2e2", alpha = 0.4)+ # spin-up edf8fb
      # annotate("rect",xmin=1861, xmax=1960, ymin=0, ymax=Inf, fill = "#66c2a4", alpha = 0.4)+ # transition b2e2e2
      annotate("rect",xmin=1961, xmax=2010, ymin=0, ymax=Inf, fill = "#238b45", alpha = 0.4)+ # projection 66c2a4
      # annotate("rect",xmin=2010, xmax=2017, ymin=0, ymax=Inf, fill = "#238b45", alpha = 0.4)+ # validation 238b45
      geom_text(data = ann_text,label = ann_text$lab, color = "Black", size = 2)+
      geom_point(data = filter(final_df_plot, Data == "Reconstructed values"),size=1)+
      geom_line(data = filter(final_df_plot, Data == "GAM predictions")) +
      scale_color_manual(values = c("blue", "black"), drop = FALSE)+
      scale_fill_chris()+
      my_theme
    
    if(var == "catch_tot"){ # different reference year are only relevent when effort reconstruction is based on catches  
      if(reference == "1950"){
        plot <- plot+
          geom_vline(xintercept=1950, linetype="dashed", color = "red", size=0.5)
      }else{
        plot <- plot+
          geom_vline(xintercept=1945, linetype="dashed", color = "red", size=0.5)+
          geom_vline(xintercept=1955, linetype="dashed", color = "red", size=0.5)
      }}
    
    # save outputs 
    return(list(reconstructed_effort_LME = final_df, plot_reconstructed_effort_LME = plot, chechDataUsed = chechDataUsed)) 
    
  # } # END if data exist 
  
} 

```

## Extrapolate total effort by LME loop

```{r extrapolate loop, echo = FALSE, fig.width=10, fig.height=7}

# SEE FIRST AND SECOND LOOP WITH FINAL DECISION ON GLOBAL FILE. 
# final loop after DECISION ----

# if catch: include war, ref 1945-1955
# if effort - it does not matter

# remove outliers first??
# Chatham.Rise.LL.NZ (but used catches for spin up) - and spin-up OK 
# Cook.Strait as above
# Prydz.Bay (yes need to remove)
# Tasman.Golden.Bays.NZ (yes need to remove)

# REMOVE OUTLIERS AS PER BELOW # WARNING to do!
explore <- effort %>%
  dplyr::filter(region == "Prydz Bay") %>%
  group_by(region, Year) %>% 
  summarise(NomActive = sum(NomActive))

max <- max(explore$NomActive)
max <- mean(explore$NomActive)+(2*sd(explore$NomActive))
explore %>% filter(NomActive >= max)

effort %>% filter(region == "Prydz Bay", Year %in% c("1978", "1979"))

# end removing outliers 


# fix the k argument in GAM (based on war_param = "include" and reference = 1950)
toKeep <- sort(unique(effort$region))

df_kappa = 6 
# with kappa = 3 bumps in effort disappear and we force the perfect logarithmic shape on the catch/effort data, but GAM in most cases does not fit the data particularly well (as does not follow bumps ...)

if(df_kappa ==6){ # better capture trends 
  
  # but avoid an increase in effort from 1950 to 1841: 
  kappa_values <- data.frame(lme = toKeep, kappa = 6)  
  # NEED TO MODIFY KAPPA IF EFFORT TRENDS NOT AS EXPECTED 
  kappa_values[c(6,8,14,17,21), 2]<-5
  kappa_values[c(6,14), 2]<-4
  
}else if (df_kappa == 3){ # assumes a "perfect" logistic 
  
  kappa_values <- data.frame(lme = toKeep, kappa = 3)

}

# # REGION needed adjusments (but see if removing outliers is ok)
# eastern bass strait 6
# eastern scotial shelf 7
# NorthernHumboldt 14
# Prydz.Bay 17
# Tasman.Golden.Bays.NZ 21

reconstructed_effort_LME_war_ref <- list()
plot_reconstructed_effort_LME_war_ref <- list()

tic()
for(i in 1:length(toKeep)){
  
  # i = 2
  print(toKeep[[i]])

  # yes war data when fitting GAM do not re-add original data as otherwise you overwrite predictions, reference mean over a decade
  b <- effort_extrapolate(toKeep[[i]], kappa = kappa_values[i,2], war_param = "include", war_decision = "no-add", reference = "1945-1955")
  
  # pdf("Output/new/LME31.pdf")
  # b$plot_reconstructed_effort_LME
  # dev.off()

  reconstructed_effort_LME_war_ref[[i]] <- b$reconstructed_effort_LME
  names(reconstructed_effort_LME_war_ref)[i] <- toKeep[i]

  plot_reconstructed_effort_LME_war_ref[[i]] <- b$plot_reconstructed_effort_LME
  names(plot_reconstructed_effort_LME_war_ref)[i] <- toKeep[i]
  
}
toc() # 176 seconds

# save and print plots ----

# check empty LME and remove from the list - it should be none
plot_reconstructed_effort_LME_war_ref <- plot_reconstructed_effort_LME_war_ref[lapply(plot_reconstructed_effort_LME_war_ref,length)>0]

if(df_kappa == 6){
  name = "Output/new/reconstructed_effort_for_spinup_tot_regional_models.pdf"
}else if (df_kappa == 3){
  name = "Output/new/reconstructed_effort_for_spinup_tot_regional_models_kappa3.pdf"
}

pdf(name, height = 8, width = 6)
marrangeGrob(grobs = plot_reconstructed_effort_LME_war_ref, nrow=2, ncol=1)
dev.off()

```

## Allocate reconstructed effort across groups function

```{r allocate, echo = FALSE}

decision <- reconstructed_effort_LME_war_ref


effort_spread <- function(historical_effort, recent_effort, toKeep, reference){
  
  # # # trial
  # historical_effort = decision[[32]] # effort from 1841 to 2017 at LME level from analysis above - missing groups (eez, gear, fgroup, saup, sector, etc.)
  # recent_effort = recent_effort # effort from 1950 to 2017 with all groups
  # toKeep = toKeep[[32]]
  # reference = "1950-1960"
  
  # calculate effort distribution across group using 1950-1960, 
  if(reference == "1950"){
    recent_effortB <- recent_effort %>% 
      filter(Year == 1950)
    }else{
      recent_effortB <- recent_effort %>% 
        filter(Year >= 1950, Year <=1960)
    }
  
  # this is done for each LME (LME in grouping below is not necessary). 
  # recent_effort is effort from 1950-2017 and recent_effortB is effort from 1950-1960 (reference years) including all groups
  # NOTE from recent_effort you consider CONTRIBUTIONS by groups 
  # first you calculate the average effort by group 1950-1960 (reference effort)
  # then you calculate the effort contribution of each group. at the LME level this sums to 1 
  
  ## NOTE: Some regions (like Ross Sea) don't have any fishing until ~1970. When this is the case, we just skip it and don't reconstruct the effort the same way, since we don't have any data to inform it. 
  
  recent_effortB <- recent_effortB %>% 
    group_by(region,SAUP,Gear,FGroup,Sector) %>% 
    summarise(NomActive = mean(NomActive, na.rm = TRUE)) %>% 
    ungroup() %>% 
    mutate(contribution = NomActive/sum(NomActive, na.rm = TRUE))
  
  # check
  sum(recent_effortB$contribution, na.rm = TRUE)
  
  # now you work on the historical data 1841-1950.
  # FIRST - create all combinations of groups in historical years - e.g. 1841 all model_area_name, saup, grear, fgroup, sector present in the reference years (1950-1960) with contribution being constant across year (proportions don't change, effort does), 1842 same story ... 
  Year = seq(1841,1949)
  toExpend <- recent_effortB %>% 
    select(-NomActive) %>% 
    unique()

  if(nrow(toExpend) == 0){
    
toExpend <- toExpend %>% 
  tibble::add_row(region = toKeep, SAUP = NA_integer_, Gear = NA_character_,
                  FGroup = NA_character_, Sector = NA_character_, contribution = NA_real_)

  toExpend <- rbindlist(lapply(1841:1949, function(x,d) data.frame(d, Year=x), d=toExpend))
  toExpend <- as_tibble(toExpend)

    
  }else{
  
  toExpend <- rbindlist(lapply(1841:1949, function(x,d) data.frame(d, Year=x), d=toExpend))
  toExpend <- as_tibble(toExpend)
  
  }
  
  # SECOND - get effort value by year for 1841-1950
  # NOTE from historical_effort you consider EFFORT VALUES by year withing LME 
  # not much done here, just data wrangling to get year, effort dataset
  historical_effort <- historical_effort %>% 
    filter(Type == "NomActive") %>% # Here can keep catches too 
    select(-group, -Considered) %>% 
    spread(Type, Value) %>% 
    filter(!is.na(NomActive)) %>% # catch data goes to 2017 but effort to 2010 - source of NAs 
    rename(NomActive_pred = NomActive) 

  # THIRD - merge historical effort values with groups from recent_effortB (through toExpand)
  # and calculate NomActive for each group based on: historical effort values*effort contribution by each group in 1950-1960 (from recent_effortB) 
  final_df <- filter(historical_effort, Year <1950) %>% # this are the only year that we need to reconstruct and merge with recent_effort
    full_join(toExpend) %>% 
    mutate(NomActive = NomActive_pred*contribution) %>% 
    select(-NomActive_pred, -contribution)
  
  # # check - ok - the sum of effort at the LME level matches .... 
  # final_df %>% filter(Year ==1841) %>% mutate(a = sum(NomActive))
  # decision[[1]] %>% filter(Year ==1841) # compare a with original data

  # FORTH - merge merge historical effort where contributions to the groups has been calculated above (1841-1950) with recent_effort (1950-2017)
  
  # tic()
  final_df <- final_df %>% 
    full_join(recent_effort) %>% 
    as.data.frame()
  # toc()
  
  # # check they give the same ouputs -  
  # trial<-final_df1 %>% 
  #   group_by(Year) %>% 
  #   summarise(effort = sum(NomActive, na.rm = TRUE)) %>%
  #   ungroup()
  # 
  # pdf("Output/new/new/trial.pdf", height = 8, width = 6)
  # ggplot(trial, aes(x = Year, y = effort))+
  #   geom_point()
  # dev.off()
  
  # final df for plotting 
  data_plot <- final_df %>%
    group_by(Year, Gear) %>% 
    summarise(NomActive = sum(NomActive, na.rm = TRUE)) %>% 
    ungroup()

  # plot <- ggplot(data_plot, aes(y = NomActive, x = Year, group = Gear, color = Gear))+
  #   geom_line()+
  #   ggtitle(paste("LME", toKeep, sep = " ")) +# WARNING - check this is correct
  #   my_theme

  # create data for annotation 
  yMax = max(data_plot$NomActive, na.rm = TRUE)
  ann_text <- data.frame(Year = c(1841+10,1961+15), NomActive = rep(yMax,2), lab = c("Transition","Experiment"), Gear = rep(unique(data_plot$Gear)[1], 2))
  
  plot <- ggplot(data = data_plot, aes(x = Year, y = NomActive, group = Gear, color = Gear)) +
    ggtitle(paste(toKeep, sep = " "))+ # WARNING - check this is correct
    annotate("rect",xmin=1841, xmax=1960, ymin=0, ymax=Inf, fill = "#b2e2e2", alpha = 0.4)+ # spin-up edf8fb
    annotate("rect",xmin=1961, xmax=2010, ymin=0, ymax=Inf, fill = "#238b45", alpha = 0.4)+ # projection 66c2a4
    geom_text(data = ann_text,label = ann_text$lab, color = "Black", size = 2)+
    geom_line(size = 0.5)+
    geom_point(size = 0.5)+
    my_theme
    
  if(reference == "1950"){
    plot <- plot+
      geom_vline(xintercept=1950, linetype="dashed", color = "red", size=0.5)
  }else{
    plot <- plot+
      geom_vline(xintercept=1950, linetype="dashed", color = "red", size=0.5)+
      geom_vline(xintercept=1960, linetype="dashed", color = "red", size=0.5)
  }
  
  # # check with gear and FGroup - OK
  # pdf("Output/new/new/trial1.pdf", height = 8, width = 6)
  # plot
  # dev.off()
    
  return(list(final_df = final_df, plot = plot))
  
}

```

## Allocate reconstructed effort across groups loop

```{r allocate loop, echo = FALSE, fig.width=10, fig.height=7}

# decide on version above: 
decision <- reconstructed_effort_LME_war_ref

# First loop ----
# SEE DECISION on GLOBAL FILE 
toKeep <- sort(unique(effort$region))
reconstructed_effort_ref <- list()
plot_reconstructed_effort_ref <- list()

for(i in 1:length(decision)){

  # i = 32
  print(toKeep [[i]])
  
  recent_effort <- filter(effort, region == toKeep [[i]])
  
  # WARNING - need to check that LMEs are matching across datasets and with toKeep - YES matching 
  b <- effort_spread(historical_effort = decision[[i]], recent_effort = recent_effort, toKeep = toKeep[[i]], reference = "1950-1960")
  
  reconstructed_effort_ref[[i]] = b$final_df
  names(reconstructed_effort_ref)[i] <- toKeep[i]
  plot_reconstructed_effort_ref[[i]] = b$plot
  names(plot_reconstructed_effort_ref)[i] <- toKeep[i]
  
}

# Decision ----

# reference mean values 1950-1960

# reconstructed_effort_ref
# plot_reconstructed_effort_ref

# save final data and plots ----

# check empty LME and remove from the list - it should be none
plot_reconstructed_effort_ref <- plot_reconstructed_effort_ref[lapply(plot_reconstructed_effort_ref,length) > 0]

pdf("Output/new/reconstructed_effort_for_spinup_regional_models.pdf", height = 8, width = 6)
marrangeGrob(grobs = plot_reconstructed_effort_ref, nrow=2, ncol=1)
dev.off()

# from list to df data ----

reconstructed_effort_ref_all <- do.call(rbind.data.frame, reconstructed_effort_ref)

```


## save final data 

```{r}
# final data adjustment
# head(reconstructed_effort_ref) # remove P, GT, NV, add spin-up, transition, experiment, validation column
reconstructed_effort_ref_all <- reconstructed_effort_ref_all %>% 
  select(-P, -NV) %>% 
  mutate(Phase = case_when (Year %in% c(1841:1860) ~ "spin-up",
                            Year %in% c(1861:1960) ~ "transition",
                            Year %in% c(1961:2010) ~ "experiment",
                            Year > 2010 ~ "validation"))

# after FishMIP meeting 
fwrite(x = reconstructed_effort_ref_all, file.path(gem_dir, "private/users/yannickr/effort_histsoc_1841_2017_regional_models.csv"))

```
